+++ pyserini/encode/_doc_encoders.py
@@ -16,10 +16,11 @@
 
 import sklearn
 import torch
+from torch.cuda.amp import autocast
 from transformers import AutoModel, AutoTokenizer, BertModel, BertTokenizer, DPRContextEncoder, \
     DPRContextEncoderTokenizer, RobertaTokenizer
 
-from pyserini.encode import AnceEncoder, DocumentEncoder
+from pyserini.encode import AnceEncoder, DocumentEncoder, UniCoilEncoder
 
 
 class TctColBertDocumentEncoder(DocumentEncoder):
@@ -29,7 +30,7 @@ class TctColBertDocumentEncoder(DocumentEncoder):
         self.model.to(self.device)
         self.tokenizer = BertTokenizer.from_pretrained(tokenizer_name or model_name)
 
-    def encode(self, texts, titles=None):
+    def encode(self, texts, titles=None, **kwargs):
         if titles is not None:
             texts = [f'[CLS] [D] {title} {text}' for title, text in zip(titles, texts)]
         else:
@@ -56,7 +57,7 @@ class DprDocumentEncoder(DocumentEncoder):
         self.model.to(self.device)
         self.tokenizer = DPRContextEncoderTokenizer.from_pretrained(tokenizer_name or model_name)
 
-    def encode(self, texts, titles=None):
+    def encode(self, texts, titles=None, **kwargs):
         max_length = 256  # hardcode for now
         if titles:
             inputs = self.tokenizer(
@@ -88,7 +89,7 @@ class AnceDocumentEncoder(DocumentEncoder):
         self.model.to(self.device)
         self.tokenizer = RobertaTokenizer.from_pretrained(tokenizer_name or model_name)
 
-    def encode(self, texts, titles=None):
+    def encode(self, texts, titles=None, **kwargs):
         if titles is not None:
             texts = [f'{title} {text}' for title, text in zip(titles, texts)]
         max_length = 512  # hardcode for now
@@ -114,7 +115,7 @@ class AutoDocumentEncoder(DocumentEncoder):
         self.pooling = pooling
         self.l2_norm = l2_norm
 
-    def encode(self, texts, titles=None):
+    def encode(self, texts, titles=None, **kwargs):
         if titles:
             texts = [f'{title} {text}' for title, text in zip(titles, texts)]
         inputs = self.tokenizer(
@@ -142,3 +143,69 @@ def mean_pooling(last_hidden_state, attention_mask):
     sum_embeddings = torch.sum(token_embeddings * input_mask_expanded, 1)
     sum_mask = torch.clamp(input_mask_expanded.sum(1), min=1e-9)
     return sum_embeddings / sum_mask
+
+
+class UniCoilDocumentEncoder(DocumentEncoder):
+    def __init__(self, model_name, tokenizer_name=None, device='cuda:0'):
+        self.device = device
+        self.model = UniCoilEncoder.from_pretrained(model_name)
+        self.model.to(self.device)
+        self.tokenizer = BertTokenizer.from_pretrained(tokenizer_name or model_name)
+
+    def encode(self, texts, titles=None, expands=None, fp16=False, **kwargs):
+        if titles:
+            texts = [f'{title} {text}' for title, text in zip(titles, texts)]
+        max_length = 512  # hardcode for now
+        if expands:
+            input_ids = self._tokenize_with_injects(texts, expands)
+        else:
+            input_ids = self.tokenizer(texts, max_length=max_length, padding='longest',
+                                       truncation=True, add_special_tokens=True,
+                                       return_tensors='pt').to(self.device)["input_ids"]
+        if fp16:
+            with autocast():
+                with torch.no_grad():
+                    batch_weights = self.model(input_ids).cpu().detach().numpy()
+        else:
+            batch_weights = self.model(input_ids).cpu().detach().numpy()
+        batch_token_ids = input_ids.cpu().detach().numpy()
+        return self._output_to_weight_dicts(batch_token_ids, batch_weights)
+
+    def _output_to_weight_dicts(self, batch_token_ids, batch_weights):
+        to_return = []
+        for i in range(len(batch_token_ids)):
+            weights = batch_weights[i].flatten()
+            tokens = self.tokenizer.convert_ids_to_tokens(batch_token_ids[i])
+            tok_weights = {}
+            for j in range(len(tokens)):
+                tok = str(tokens[j])
+                weight = float(weights[j])
+                if tok == '[CLS]':
+                    continue
+                if tok == '[PAD]':
+                    break
+                if tok not in tok_weights:
+                    tok_weights[tok] = weight
+                elif weight > tok_weights[tok]:
+                    tok_weights[tok] = weight
+            to_return.append(tok_weights)
+        return to_return
+
+    def _tokenize_with_injects(self, texts, expands):
+        tokenized = []
+        max_len = 0
+        for text, expand in zip(texts, expands):
+            text_ids = self.tokenizer.encode(text, add_special_tokens=False, max_length=400, truncation=True)
+            expand_ids = self.tokenizer.encode(expand, add_special_tokens=False, max_length=100, truncation=True)
+            injects = set()
+            for tok_id in expand_ids:
+                if tok_id not in text_ids:
+                    injects.add(tok_id)
+            all_tok_ids = [101] + text_ids + [102] + list(injects) + [102]  # 101: CLS, 102: SEP
+            tokenized.append(all_tok_ids)
+            cur_len = len(all_tok_ids)
+            if cur_len > max_len:
+                max_len = cur_len
+        for i in range(len(tokenized)):
+            tokenized[i] += [0] * (max_len - len(tokenized[i]))
+        return torch.tensor(tokenized, device=self.device)
